feat(core) astubbs#255: the rider is charged against the metadata cap, never against back-pressure, and the ladder never costs the hole map
Back-pressure exists so a partition's metadata can shrink as work completes, and rider bytes do
not shrink. Charged against the 0.75 threshold a rider is a floor the mechanism cannot relieve, and
on a caught-up partition a permanent block, because nothing is in flight to complete. So
updateBlockFromEncodingResult now takes two lengths: the hole encoding's own character length is
judged against the threshold, and the assembled string's against the cap. With no rider configured
the first number is the assembled string's own length, byte for byte, which is why the existing
back-pressure test's block point does not move - pinned over a corpus of hole densities by
recomputing the old rule independently.

When the assembled payload would exceed the cap, a ladder chooses what to shed, predicted from the
Base64 closed form so the outer codec runs once per commit on the winning rung: the rider, then the
three-byte drop marker, then the envelope altogether - because a hole map that sits within the
marker's cost of the cap must still commit, and configuring a rider may never cost a partition
metadata it would otherwise have written. Only after all three does today's strip fire. A rider the
supplier's own cap already accepted provably fits alongside its offset map, so in practice the
descent starts from the marker; the rung chooser is tested on its own boundary table regardless.

The rung enum lives in its own public file only because the Truth subject generator recurses into
every type in the state package and cannot see a package-private one.

Also fixes a vacuous pair of assertions in the previous unit's guard test: comparing a decoded byte
to the magic-byte constant boxed one side to Integer and the other to Byte, so the inequality could
never fail. Both sides are ints now.

Test-first: two scenarios went red on today's behaviour - a rider over the threshold blocked the
partition, and a hole map within the marker's cost of the cap was stripped outright - then green
with the whole core unit suite.

Plan: docs/plans/2026-09-05-001-feat-offset-metadata-rider-plan.md, U2.

+++ parallel-consumer-core/src/test/java/bz/stub/parallelconsumer/state/RiderSupplierGuardTest.java
@@ -241,7 +241,7 @@ class RiderSupplierGuardTest {
         assertWithMessage("a throwing supplier must cost the rider only - no envelope is written, so the commit "
                         + "carries exactly what this build commits with no supplier configured at all")
                 .that(Base64.getDecoder().decode(first.metadata())[0])
-                .isNotEqualTo(OffsetRiderEnvelope.MAGIC_BYTE);
+                .isNotEqualTo((int) OffsetRiderEnvelope.MAGIC_BYTE); // int vs int: Integer.equals(Byte) is never true
         assertWithMessage("R15: the failure is logged and treated as no rider, so the NEXT commit still happens")
                 .that(second.offset())
                 .isAtLeast(first.offset());
@@ -487,7 +487,7 @@ class RiderSupplierGuardTest {
                         + "envelope with an empty rider in it would still decode, which is exactly what this "
                         + "distinguishes", whatItDid)
                 .that(payload[0])
-                .isNotEqualTo(OffsetRiderEnvelope.MAGIC_BYTE);
+                .isNotEqualTo((int) OffsetRiderEnvelope.MAGIC_BYTE); // int vs int: Integer.equals(Byte) is never true
         assertWithMessage("and the hole map the partition would have committed anyway is untouched")
                 .that(OffsetMapCodecManager
                         .deserialiseIncompleteOffsetMapFromBase64(committed.offset(), committed.metadata())
